♻️ Separate Manifest and LocalMod
