+++ src/core.rs
@@ -1,5 +1,7 @@
 pub mod everest;
 pub mod loader;
+pub mod local;
 pub mod network;
+pub mod registry;
 pub mod resolver;
 pub mod update;
