+++ src/main.rs
@@ -14,8 +14,8 @@ mod commands;
 mod config;
 mod core;
 mod dependency;
-mod local_mods;
 mod log;
+mod manifest;
 mod mirror;
 mod registry;
 mod ui;
